Finished User (1) Edit Password, (2) Sign Out, (3) Build Profile.
-sign-in now makes API request for user profile; if user has profile it
brings you right to chat view. The Log out button in the nav bar and the
edit password button are both fully functional.

-if the user does not have a profile, a create profile modal pops up.
This modal has front-end JS (see bootstrap js file) that auto populates
it in a 999-999-9999 format, which is how the back-end validator
expects it. Users cannot enter letters or more than 10 digits. If a user
enters an insufficient # of digits the back-end validator is triggered
and the client provides an appropriate notification.

***NEXT***
(1) Finish Profile CRUD
(2) THEN move on to building chat functionality